+++ include/vsg/core/Array3D.h
@@ -72,9 +72,9 @@ namespace vsg
             size_t original_size = size();
 
             Data::read(input);
-            size_t width = input.readValue<uint32_t>("Width");
-            size_t height = input.readValue<uint32_t>("Height");
-            size_t depth = input.readValue<uint32_t>("Depth");
+            uint32_t width = input.readValue<uint32_t>("Width");
+            uint32_t height = input.readValue<uint32_t>("Height");
+            uint32_t depth = input.readValue<uint32_t>("Depth");
             size_t new_size = computeValueCountIncludingMipmaps(width, height, depth, _layout.maxNumMipmaps);
             if (input.matchPropertyName("Data"))
             {
